+++ examples/iris.py
@@ -1,0 +1,61 @@
+from PyNomaly import loop
+import pandas as pd
+from pydataset import data
+from sklearn.cluster import DBSCAN
+import matplotlib.pyplot as plt
+from mpl_toolkits.mplot3d import Axes3D
+
+
+iris = pd.DataFrame(data('iris'))
+iris = pd.DataFrame(iris.drop('Species', 1))
+
+
+db = DBSCAN(eps=0.9, min_samples=10).fit(iris)
+m = loop.LocalOutlierProbability(iris).fit()
+scores_noclust = m.local_outlier_probabilities
+m_clust = loop.LocalOutlierProbability(iris, cluster_labels=list(db.labels_)).fit()
+scores_clust = m_clust.local_outlier_probabilities
+
+
+iris_clust = pd.DataFrame(iris.copy())
+iris_clust['scores'] = scores_clust
+iris_clust['labels'] = db.labels_
+
+iris['scores'] = scores_noclust
+
+
+fig = plt.figure(figsize=(7, 7))
+ax = fig.add_subplot(111, projection='3d')
+ax.scatter(iris['Sepal.Width'], iris['Petal.Width'], iris['Sepal.Length'],
+c=iris['scores'], cmap='seismic', s=50)
+ax.set_xlabel('Sepal.Width')
+ax.set_ylabel('Petal.Width')
+ax.set_zlabel('Sepal.Length')
+plt.show()
+plt.clf()
+plt.cla()
+plt.close()
+
+fig = plt.figure(figsize=(7, 7))
+ax = fig.add_subplot(111, projection='3d')
+ax.scatter(iris_clust['Sepal.Width'], iris_clust['Petal.Width'], iris_clust['Sepal.Length'],
+c=iris_clust['scores'], cmap='seismic', s=50)
+ax.set_xlabel('Sepal.Width')
+ax.set_ylabel('Petal.Width')
+ax.set_zlabel('Sepal.Length')
+plt.show()
+plt.clf()
+plt.cla()
+plt.close()
+
+fig = plt.figure(figsize=(7, 7))
+ax = fig.add_subplot(111, projection='3d')
+ax.scatter(iris_clust['Sepal.Width'], iris_clust['Petal.Width'], iris_clust['Sepal.Length'],
+c=iris_clust['labels'], cmap='Set1', s=50)
+ax.set_xlabel('Sepal.Width')
+ax.set_ylabel('Petal.Width')
+ax.set_zlabel('Sepal.Length')
+plt.show()
+plt.clf()
+plt.cla()
+plt.close()